refactor: extract Draggable component
