+++ docs/log/2025-08-26.md
@@ -24,5 +24,6 @@
 - Added numeric X/Y inputs to LogoPanel that mirror drag movements.
 - Hoisted draggable component to preserve state and allow continuous logo dragging.
 - Clamped draggable elements using offset dimensions to prevent edge clipping on all sides.
+- Extracted Draggable component into its own module and added unit tests.
